xmls and java files for week7 updated

+++ mobile/src/main/java/camera/Capture.java
@@ -2,7 +2,8 @@ package camera;
 
 public interface Capture {
 	public Integer mg_pixels();
-	public Integer ap();
+
+	public Double ap();
 	public Boolean OIS();
 
 }
